- Riorganizzati i file e gli include: separate dichiarazioni e definizioni;
