+++ OldGame01_Maker/OldGame01_Maker.vcxproj
@@ -11,7 +11,22 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="..\src\ClLevel.cpp" />
+    <ClCompile Include="..\src\clmap.cpp" />
+    <ClCompile Include="..\src\cltexture.cpp" />
+    <ClCompile Include="..\src\Common.cpp" />
+    <ClCompile Include="..\src\GmMakDef.cpp" />
+    <ClCompile Include="..\src\GmMakFun.cpp" />
     <ClCompile Include="..\src\LvlMaker.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\src\cllevel.h" />
+    <ClInclude Include="..\src\ClMap.h" />
+    <ClInclude Include="..\src\ClTexture.h" />
+    <ClInclude Include="..\src\Common.h" />
+    <ClInclude Include="..\src\Costanti.h" />
+    <ClInclude Include="..\src\GmMakDef.h" />
+    <ClInclude Include="..\src\GmMakFun.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{2AB51858-242E-47D5-8A80-57C810FC27B6}</ProjectGuid>
@@ -66,6 +81,7 @@
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
+      <AdditionalDependencies>allegro-4.4.2-mt.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
@@ -84,6 +100,7 @@
       <GenerateDebugInformation>true</GenerateDebugInformation>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
+      <AdditionalDependencies>allegro-4.4.2-mt.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
