+++ src/Common.cpp
@@ -1,4 +1,5 @@
 
+#include "Common.h"
 
 FILE *filelog;
