+++ src/Engine.h
@@ -1,0 +1,14 @@
+
+#ifndef _ENGINE_H_
+#define _ENGINE_H_
+
+extern bool button[128];	//	flag che servono per evitare la ripetizione
+							//	automatica dei tasti;
+
+/************************************************************************
+*							Programma  Principale  						*
+************************************************************************/
+
+void engine(void);
+
+#endif
